django-nextjs-backend-api/requirements.txt
Django
django-cors-headers
django-ninja==1.4.3
django-ninja-jwt[crypto]==5.3.7
dj-database-url
psycopg[binary] # psycopg
pydantic[email]
python-decouple
rav